CosmoMMF

CosmoMMF.kspace_gaussian_filter

CosmoMMF.smooth_loggauss

CosmoMMF.wavevectors3D

CosmoMMF.xθ

CosmoMMF.θ

CosmoMMF.kspace_gaussian_filter — Method

kspace_gaussian_filter(R_S, kv)

Computes the Fourier representation of a Gaussian filter with width $R_S$. You would multiply this with the FFT'd field and then IFFT to apply the filter.

Arguments:

R_S::T: filter radius

kv: wavevector, tuple of (kx, ky, kz)

Returns:

Array{T,3}: fourier representation of gaussian filter

CosmoMMF.smooth_loggauss — Method

kspace_gaussian_filter(f, R_S, kv)

Computes the Fourier representation of a Gaussian filter with width $R_S$. You would multiply this with the FFT'd field and then IFFT to apply the filter.

Arguments:

f: the 3D array to smooth

R_S::T: filter radius

kv: wavevector, tuple of (kx, ky, kz)

Returns:

Array{T,3}: fourier representation of gaussian filter

CosmoMMF.wavevectors3D — Function

wavevectors3D([T=Float64], dims, box_size=(2π, 2π, 2π))

Arguments:

T : output element type

dims: array or tuple containing size

box_size: side lengths of box, normalizes the wavenumber

Returns:

Tuple{Vector{T},Vector{T},Vector{T}}: kx, ky, kz wavevectors

Examples

julia> CosmoMMF.wavevectors3D((6,2,2))
([0.0, 1.0, 2.0, -3.0, -2.0, -1.0], [0.0, -1.0], [0.0, -1.0])

CosmoMMF.xθ — Method

xθ(x)

Heaviside step times x. Returns x if the argument is positive, zero otherwise.

CosmoMMF.θ — Method

θ(x)

Heaviside step function. Returns one if the argument is positive, zero otherwise.